refactor: 상품 디테일 관련 타입 추가
