+++ src/interface/cart_I.ts
@@ -1,0 +1,9 @@
+export interface cartItem {
+  id: string;
+  productCount: number;
+  productImage: string;
+  productName: string;
+  productPrice: number;
+  userImage: string;
+  userName: string;
+}
